chore: reorganize thinking tests
